review comments, moved functions to new file
